+++ src/common/VSCServerManagerBlob.ts
@@ -87,7 +87,8 @@ export default class VSCServerManagerBlob extends VSCServerManagerBase {
       env.key(),
       env.pwd(),
       env.oauth(),
-      env.disableProductStyleUrl()
+      env.disableProductStyleUrl(),
+      env.inMemoryPersistence(),
     );
     return config;
   }