+++ src/common/VSCServerManagerClosedState.ts
@@ -1,4 +1,6 @@
+import { setExtentMemoryLimit } from "./ConfigurationBase";
 import IVSCServerManagerState from "./IVSCServerManagerState";
+import VSCEnvironment from "./VSCEnvironment";
 import VSCServerManagerBase from "./VSCServerManagerBase";
 import { VSCServerManagerRunningState } from "./VSCServerManagerRunningState";
 
@@ -7,8 +9,12 @@ export default class VSCServerManagerClosedState
   public async start(
     serverManager: VSCServerManagerBase
   ): Promise<IVSCServerManagerState> {
+    const env = new VSCEnvironment()
+    setExtentMemoryLimit(env, false)
+
     await serverManager.createImpl();
     await serverManager.startImpl();
+
     return new VSCServerManagerRunningState();
   }