+++ src/queue/QueueConfiguration.ts
@@ -1,5 +1,6 @@
 import ConfigurationBase from "../common/ConfigurationBase";
 import { StoreDestinationArray } from "../common/persistence/IExtentStore";
+import { MemoryExtentChunkStore } from "../common/persistence/MemoryExtentStore";
 import {
   DEFAULT_ENABLE_ACCESS_LOG,
   DEFAULT_ENABLE_DEBUG_LOG,
@@ -39,7 +40,9 @@ export default class QueueConfiguration extends ConfigurationBase {
     key: string = "",
     pwd: string = "",
     oauth?: string,
-    disableProductStyleUrl: boolean = false
+    disableProductStyleUrl: boolean = false,
+    public readonly isMemoryPersistence: boolean = false,
+    public readonly memoryStore?: MemoryExtentChunkStore,
   ) {
     super(
       host,