+++ src/queue/main.ts
@@ -12,6 +12,7 @@ import {
   DEFAULT_QUEUE_PERSISTENCE_ARRAY,
   DEFAULT_QUEUE_PERSISTENCE_PATH
 } from "./utils/constants";
+import { setExtentMemoryLimit } from "../common/ConfigurationBase";
 
 // tslint:disable:no-console
 
@@ -64,7 +65,8 @@ async function main() {
     env.key(),
     env.pwd(),
     env.oauth(),
-    env.disableProductStyleUrl()
+    env.disableProductStyleUrl(),
+    env.inMemoryPersistence(),
   );
 
   // We use logger singleton as global debugger logger to track detailed outputs cross layers
@@ -75,6 +77,8 @@ async function main() {
 
   // Create server instance
   const server = new QueueServer(config);
+
+  setExtentMemoryLimit(env, true);
 
   // Start server
   console.log(