+++ src/table/TableConfiguration.ts
@@ -35,7 +35,8 @@ export default class TableConfiguration extends ConfigurationBase {
     key: string = "",
     pwd: string = "",
     oauth?: string,
-    disableProductStyleUrl: boolean = false
+    disableProductStyleUrl: boolean = false,
+    public readonly isMemoryPersistence: boolean = false,
   ) {
     super(
       host,