+++ src/table/main.ts
@@ -29,7 +29,7 @@ async function main() {
     await access(dirname(debugFilePath));
   }
 
-  // Store table configuation
+  // Store table configuration
   const config = new TableConfiguration(
     env.tableHost(),
     env.tablePort(),
@@ -44,7 +44,8 @@ async function main() {
     env.key(),
     env.pwd(),
     env.oauth(),
-    env.disableProductStyleUrl()
+    env.disableProductStyleUrl(),
+    env.inMemoryPersistence(),
   );
 
   // We use logger singleton as global debugger logger to track detailed outputs cross layers
@@ -78,7 +79,7 @@ async function main() {
     })
     .once("SIGINT", () => {
       console.log(beforeCloseMessage);
-      server.clean().then(() => {
+      server.close().then(() => {
         console.log(afterCloseMessage);
       });
     });