+++ tests/queue/apis/messageid.test.ts
@@ -9,7 +9,6 @@ import {
 
 import { configLogger } from "../../../src/common/Logger";
 import { StoreDestinationArray } from "../../../src/common/persistence/IExtentStore";
-import QueueConfiguration from "../../../src/queue/QueueConfiguration";
 import Server from "../../../src/queue/QueueServer";
 import {
   EMULATOR_ACCOUNT_KEY,
@@ -18,6 +17,7 @@ import {
   rmRecursive,
   sleep
 } from "../../testutils";
+import QueueTestServerFactory from "../utils/QueueTestServerFactory";
 
 // Set true to enable debug log
 configLogger(false);
@@ -30,22 +30,13 @@ describe("MessageId APIs test", () => {
   const extentDbPath = "__extentTestsStorage__";
   const persistencePath = "__queueTestsPersistence__";
 
-  const DEFUALT_QUEUE_PERSISTENCE_ARRAY: StoreDestinationArray = [
+  const DEFAULT_QUEUE_PERSISTENCE_ARRAY: StoreDestinationArray = [
     {
       locationId: "queueTest",
       locationPath: persistencePath,
       maxConcurrency: 10
     }
   ];
-
-  const config = new QueueConfiguration(
-    host,
-    port,
-    metadataDbPath,
-    extentDbPath,
-    DEFUALT_QUEUE_PERSISTENCE_ARRAY,
-    false
-  );
 
   const baseURL = `http://${host}:${port}/devstoreaccount1`;
   const serviceClient = new QueueServiceClient(
@@ -68,7 +59,11 @@ describe("MessageId APIs test", () => {
   const messageContent = "Hello World";
 
   before(async () => {
-    server = new Server(config);
+    server = new QueueTestServerFactory().createServer({
+      metadataDBPath: metadataDbPath,
+      extentDBPath: extentDbPath,
+      persistencePathArray: DEFAULT_QUEUE_PERSISTENCE_ARRAY,
+    });
     await server.start();
   });