+++ tests/table/apis/table.validation.rest.test.ts
@@ -5,7 +5,6 @@
 // special care is needed to replace etags and folders when used
 import * as assert from "assert";
 import { configLogger } from "../../../src/common/Logger";
-import TableConfiguration from "../../../src/table/TableConfiguration";
 import TableServer from "../../../src/table/TableServer";
 import {  getUniqueName } from "../../testutils";
 import {
@@ -16,25 +15,15 @@ import {
   getToAzuriteProductionUrl
 } from "../utils/table.entity.tests.rest.submitter";
 import dns = require("dns");
+import TableTestServerFactory from "../utils/TableTestServerFactory";
 
 // Set true to enable debug log
 configLogger(false);
 
 describe("table name validation tests", () => {
-  const host = "127.0.0.1";
-  const port = 11002;
   const metadataDbPath = getUniqueName("__tableTestsStorage__");
   const enableDebugLog: boolean = true;
   const debugLogPath: string = "g:/debug.log";
-  const config = new TableConfiguration(
-    host,
-    port,
-    metadataDbPath,
-    enableDebugLog,
-    false,
-    undefined,
-    debugLogPath
-  );
   const productionStyleHostName = "devstoreaccount1.table.localhost"; // Use hosts file to make this resolve
   const productionStyleHostNameForSecondary = "devstoreaccount1-secondary.table.localhost";
 
@@ -43,7 +32,14 @@ describe("table name validation tests", () => {
   let tableName: string = getUniqueName("flows");
 
   before(async () => {
-    server = new TableServer(config);
+    server = new TableTestServerFactory().createServer({
+      metadataDBPath: metadataDbPath,
+      enableDebugLog: enableDebugLog,
+      debugLogFilePath: debugLogPath,
+      loose: false,
+      skipApiVersionCheck: false,
+      https: false
+    });
     await server.start();
   });
 
@@ -196,8 +192,12 @@ describe("table name validation tests", () => {
   });
 
   it("should create a table with a name which is a substring of an existing table, @loki", async () => {
-    tableName = getUniqueName("table");
-    const body = JSON.stringify({
+    // this will be used for 3 table names
+    // [ a + b ], [ a ], [ b ]
+    const a = getUniqueName("t")
+    const b = getUniqueName("t")
+    tableName = a + b;
+    const body1 = JSON.stringify({
       TableName: tableName
     });
     const createTableHeaders = {
@@ -205,7 +205,7 @@ describe("table name validation tests", () => {
       Accept: "application/json;odata=nometadata"
     };
     try {
-      await postToAzurite("Tables", body, createTableHeaders);
+      await postToAzurite("Tables", body1, createTableHeaders);
     } catch (err: any) {
       assert.strictEqual(
         err.response.status,
@@ -213,9 +213,8 @@ describe("table name validation tests", () => {
         `unexpected status code : ${err.response.status}`
       );
     }
-    const tableName2 = tableName.substring(0, tableName.length - 4);
     const body2 = JSON.stringify({
-      TableName: tableName2
+      TableName: a
     });
     try {
       await postToAzurite("Tables", body2, createTableHeaders);
@@ -226,9 +225,8 @@ describe("table name validation tests", () => {
         `unexpected exception with end trimmed! : ${err}`
       );
     }
-    const tableName3 = tableName.substring(4);
     const body3 = JSON.stringify({
-      TableName: tableName3
+      TableName: b
     });
     try {
       await postToAzurite("Tables", body3, createTableHeaders);
@@ -242,10 +240,14 @@ describe("table name validation tests", () => {
   });
 
   it("should delete a table with a name which is a substring of an existing table, @loki", async () => {
-    tableName = getUniqueName("table");
-    const shortName = tableName.substring(0, 6);
-    const basicBody = JSON.stringify({
-      TableName: shortName
+    // this will be used for 4 table names
+    // [ a ], [ a + b + c ], [ a + b ], [ b + c ]
+    const a = getUniqueName("t")
+    const b = getUniqueName("t")
+    const c = getUniqueName("t")
+
+    const body1 = JSON.stringify({
+      TableName: a
     });
     const createTableHeaders = {
       "Content-Type": "application/json",
@@ -259,7 +261,7 @@ describe("table name validation tests", () => {
     try {
       const createTable1 = await postToAzurite(
         "Tables",
-        basicBody,
+        body1,
         createTableHeaders
       );
       assert.strictEqual(
@@ -274,14 +276,14 @@ describe("table name validation tests", () => {
         `unexpected status code creating first table : ${err.response.status}`
       );
     }
-    const body = JSON.stringify({
-      TableName: tableName
+    const body2 = JSON.stringify({
+      TableName: a + b + c
     });
     // create table 2
     try {
       const createTable2 = await postToAzurite(
         "Tables",
-        body,
+        body2,
         createTableHeaders
       );
       assert.strictEqual(
@@ -296,15 +298,14 @@ describe("table name validation tests", () => {
         `unexpected status code : ${err.response.status}`
       );
     }
-    const tableName2 = tableName.substring(0, tableName.length - 4);
-    const body2 = JSON.stringify({
-      TableName: tableName2
+    const body3 = JSON.stringify({
+      TableName: a + b
     });
     // create table 3
     try {
       const createTable3 = await postToAzurite(
         "Tables",
-        body2,
+        body3,
         createTableHeaders
       );
       assert.strictEqual(
@@ -319,15 +320,14 @@ describe("table name validation tests", () => {
         `unexpected exception with end trimmed! : ${err}`
       );
     }
-    const tableName3 = tableName.substring(4);
-    const body3 = JSON.stringify({
-      TableName: tableName3
+    const body4 = JSON.stringify({
+      TableName: b + c
     });
     // create table 4
     try {
       const createTable4 = await postToAzurite(
         "Tables",
-        body3,
+        body4,
         createTableHeaders
       );
       assert.strictEqual(
@@ -342,7 +342,7 @@ describe("table name validation tests", () => {
         `unexpected exception with start trimmed! : ${err}`
       );
     }
-    // now list tables after deletion...
+    // now list tables before deletion...
     try {
       const listTableResult1 = await getToAzurite("Tables", createTableHeaders);
       // we count all tables created in the tests before this one as well
@@ -361,7 +361,7 @@ describe("table name validation tests", () => {
     // now delete "table"
     try {
       const deleteResult = await deleteToAzurite(
-        `Tables('${shortName}')`,
+        `Tables('${a}')`,
         "",
         createTableHeaders
       );
@@ -388,7 +388,7 @@ describe("table name validation tests", () => {
       for (const table of listTableResult2.data.value) {
         assert.notStrictEqual(
           table.TableName,
-          shortName,
+          a,
           "We still list the table we should have deleted."
         );
       }